% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTTRatesHistory.R
\name{ttInitialize}
\alias{ttInitialize}
\title{Set the server from which the info will be read}
\usage{
ttInitialize(serverName = "cryptottlivewebapi.xbtce.net",
  port = "8443")
}
\arguments{
\item{serverName}{a character. Server}

\item{port}{a numeric. Port number}
}
\description{
Set the server from which the info will be read
}
\examples{
connection <- ttInitialize(serverName = "cryptottlivewebapi.xbtce.net", port = "8443")

symbols <- connection$GetSymbolsInfo()

bars <- connection$GetBarsHistrory("BTCUSD", barsType = "Bid", periodicity = "M1", as.POSIXct("2018-10-12 00:00:00", tz = "GMT"), as.POSIXct("2018-10-12 05:00:00", tz = "GMT"), count = 0)

ticks <- connection$GetTickHistory("BTCUSD", as.POSIXct("2018-10-12 00:00:00", tz = "GMT"), as.POSIXct("2018-10-12 05:00:00", tz = "GMT"), count = 0)

currentQuotes <- connection$GetCurrentQuotes()
}
